+++ composer.json
@@ -23,7 +23,7 @@
     "minimum-stability": "dev",
     "autoload": {
         "psr-0": {
-            "fdisotto": "src/"
+            "partitaiva": "src/"
         }
     }
 }
