+++ src/partitaiva/VatService.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace fdisotto;
+namespace partitaiva;
 
 use SoapClient;
